initial refactor produce records
